libweston: fbdev: Attempt to detect the first framebuffer device in the seat. instead of defaulting to /dev/fb0
This adds a function to detect the first framebuffer device in the
current seat. Instead of hardcoding /dev/fb0, detect the device
with udev, favoring the boot_vga device, and falling back to the
first framebuffer device in the seat if there is none. This is very
similar to what compositor-drm does to find display devices